feat: refactor policy
