+++ src/components/PageList/Actions/ActionButton.vue
@@ -1,4 +1,5 @@
 <script>
+import { h } from 'vue'
 import * as R from 'ramda'
 import { hasPermission } from '@/utils/auth'
 
@@ -85,7 +86,7 @@ export default {
       action = <a-tooltip title={ this.tooltip } placement='left' destroyTooltipOnHide>{ action }</a-tooltip>
     }
     if (this.item.render) {
-      action = this.item.render(this.row, { style: this.buttonStyle, block: this.buttonClock, size: this.buttonSize, type: this.meta.buttonType || this.buttonType, disabled: this.disabled, onClick: this.handleClick })
+      action = this.item.render(this.row, { style: this.buttonStyle, block: this.buttonClock, size: this.buttonSize, type: this.meta.buttonType || this.buttonType, disabled: this.disabled, onClick: this.handleClick }, h)
     }
     v.push(action)
     return v
