+++ src/constants/policy/policy-res-list.js
@@ -1,0 +1,6 @@
+import Compute from './compute'
+
+// 定义资源数据源
+export const POLICY_RES_LIST = [
+  Compute,
+]
